requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
prompt-toolkit==1.0.14
psycopg2-binary==2.9.3
pycrypto==2.6.1
pycryptodome==3.14.1
Pygments==2.11.2
PyInquirer==1.0.3
regex==2022.3.15
requests==2.27.1
six==1.16.0
urllib3==1.26.9
wcwidth==0.2.5
